video 7

+++ src/Main/GameObject.java
@@ -7,11 +7,11 @@ import java.awt.*;
  */
 public abstract class GameObject {
 
-    protected int x,y;
+    protected float x,y;
     protected ID id;
-    protected int velX, velY;
+    protected float velX, velY;
 
-    public GameObject(int x, int y, ID id){
+    public GameObject(float x, float y, ID id){
         this.x=x;
         this.y=y;
         this.id=id;
@@ -23,11 +23,11 @@ public abstract class GameObject {
     
     public abstract Rectangle getBounds();
 
-    public int getX(){
+    public float getX(){
         return this.x;
     }
 
-    public int getY(){
+    public float getY(){
         return this.y;
     }
 
@@ -35,11 +35,11 @@ public abstract class GameObject {
         return this.id;
     }
 
-    public int getVelX(){
+    public float getVelX(){
         return this.velX;
     }
 
-    public int getvelY(){
+    public float getvelY(){
         return this.velY;
     }
